+++ praktikumsplaner-frontend/frontend/src/components/Assignment/PraktikumsstelleCard.vue
@@ -1,7 +1,30 @@
 <template>
-    <v-container>
+    <v-container
+        @drop="drop($event, value)"
+        @dragover.prevent
+        @dragenter.prevent
+    >
+        <yes-no-dialog-without-activator
+            v-model="warningDialog"
+            :dialogtitle="warningDialogTitle"
+            :dialogtext="warningDialogText"
+            @no="resetWarningDialog"
+            @yes="assignNwk"
+        ></yes-no-dialog-without-activator>
+
+        <yes-no-dialog-without-activator
+            v-model="unassignConfirmDialog"
+            :dialogtitle="unassignDialogTitle"
+            :dialogtext="unassignDialogContent"
+            @no="resetUnassign"
+            @yes="unassignNwk"
+        ></yes-no-dialog-without-activator>
         <v-card
             class="full-width-card"
+            :class="{
+                'custom-card-active': props.value.assignedNwk,
+                spacer: true,
+            }"
             elevation="16"
             outlined
             :ripple="false"
@@ -10,32 +33,29 @@
             <v-row>
                 <v-col cols="9">
                     <v-card-title
-                        >Stelle bei
-                        {{ props.praktikumsstelle.dienststelle }}</v-card-title
+                        >Stelle bei {{ props.value.dienststelle }}</v-card-title
                     >
                 </v-col>
                 <v-col>
-                    <v-card-text
-                        v-if="props.praktikumsstelle.planstelleVorhanden"
-                    >
+                    <v-card-text v-if="props.value.planstelleVorhanden">
                         <v-icon x-large>mdi-account-star</v-icon>
                     </v-card-text>
                 </v-col>
             </v-row>
             <v-card-text>
                 <p style="white-space: pre-line">
-                    {{ getCardText(props.praktikumsstelle) }}
+                    {{ getCardText(props.value) }}
                 </p></v-card-text
             >
             <v-col>
                 <v-chip
-                    v-if="props.praktikumsstelle.assignedNwk"
+                    v-if="assignedNwk"
                     color="primary"
                     close
                     close-icon="mdi-close"
-                    @click:close="openConfirmationDialog(praktikumsstelle)"
+                    @click:close="openConfirmationDialog(value)"
                     >{{
-                        `${props.praktikumsstelle.assignedNwk.vorname} ${props.praktikumsstelle.assignedNwk.nachname}`
+                        `${props.value.assignedNwk.vorname} ${props.value.assignedNwk.nachname}`
                     }}</v-chip
                 ></v-col
             >
@@ -56,18 +76,11 @@
                     <v-divider></v-divider>
                     <v-card-text>
                         <p style="white-space: pre-line">
-                            {{ getCardDetailText(props.praktikumsstelle) }}
+                            {{ getCardDetailText(props.value) }}
                         </p>
                     </v-card-text>
                 </div>
             </v-expand-transition>
-            <yes-no-dialog-without-activator
-                v-model="unassignConfirmDialog"
-                :dialogtitle="unassignDialogTitle"
-                :dialogtext="unassignDialogContent"
-                @no="resetUnassign"
-                @yes="unassignNwk"
-            ></yes-no-dialog-without-activator>
         </v-card>
     </v-container>
 </template>
@@ -77,18 +90,28 @@ import Praktikumsstelle from "@/types/Praktikumsstelle";
 import PraktikumsstellenService from "@/api/PraktikumsstellenService";
 import { EventBus } from "@/stores/event-bus";
 import YesNoDialogWithoutActivator from "@/components/common/YesNoDialogWithoutActivator.vue";
-
-const show = ref<boolean>(false);
+import Nwk from "@/types/Nwk";
 
 const props = defineProps<{
-    praktikumsstelle: Praktikumsstelle;
+    value: Praktikumsstelle;
 }>();
 
+const emits = defineEmits<{
+    (e: "input", praktikumsstelle: Praktikumsstelle): void;
+}>();
+
+const show = ref<boolean>(false);
 const unassignDialogContent = ref<string>("");
 const unassignDialogTitle = ref<string>("Zuweisung aufheben?");
 const unassignConfirmDialog = ref<boolean>(false);
-
+const warningDialog = ref<boolean>(false);
 const stelleToAssignUnassign = ref<Praktikumsstelle>();
+const nwkToAssignUnassing = ref<Nwk>();
+const warningDialogTitle = ref<string>(
+    "Warnung. Wollen sie wirklich fortfahren?"
+);
+const warningDialogText = ref<string>("");
+const assignedNwk = ref(props.value.assignedNwk);
 
 function getCardText(stelle: Praktikumsstelle): string {
     let cardText = "";
@@ -154,6 +177,192 @@ function getCardDetailText(stelle: Praktikumsstelle): string {
         stelle.taetigkeiten;
     return cardText;
 }
+function drop(event: DragEvent, stelle: Praktikumsstelle) {
+    try {
+        console.log("Drop registriert.");
+        const draggedNwk: Nwk = JSON.parse(
+            event.dataTransfer?.getData("application/json") as string
+        );
+        nwkToAssignUnassing.value = new Nwk(
+            draggedNwk.id,
+            draggedNwk.vorname,
+            draggedNwk.nachname,
+            draggedNwk.jahrgang,
+            draggedNwk.vorlesungstage,
+            draggedNwk.isActive,
+            draggedNwk.studiengang,
+            draggedNwk.ausbildungsrichtung
+        );
+    } catch (e) {
+        return;
+    }
+
+    if (
+        !stelle ||
+        !stelle.id ||
+        stelle.assignedNwk ||
+        nwkToAssignUnassing.value.id === ""
+    ) {
+        return;
+    }
+
+    // Check if Studiums or Ausbildungspraktikumsstelle
+    if (
+        stelle.ausbildungsrichtung == undefined &&
+        nwkToAssignUnassing.value.studiengang == undefined
+    ) {
+        warningDialogText.value +=
+            "Wollen sie wirklich " +
+            nwkToAssignUnassing.value.vorname +
+            " " +
+            nwkToAssignUnassing.value.nachname +
+            " auf eine Studiumspraktikumsstelle setzen, obwohl er/sie Auszubildende/r ist?\n";
+    }
+
+    if (
+        stelle.studiengang == undefined &&
+        nwkToAssignUnassing.value.ausbildungsrichtung == undefined
+    ) {
+        warningDialogText.value +=
+            "Wollen sie wirklich " +
+            nwkToAssignUnassing.value.vorname +
+            " " +
+            nwkToAssignUnassing.value.nachname +
+            " auf eine Ausbildungspraktikumsstelle setzen, obwohl er/sie Student*in ist?\n";
+    }
+
+    // Check if studiengang is the same
+    if (
+        stelle.studiengang &&
+        nwkToAssignUnassing.value.studiengang != "FISI" &&
+        stelle.studiengang != nwkToAssignUnassing.value.studiengang
+    ) {
+        warningDialogText.value +=
+            "Wollen sie wirklich eine/n " +
+            nwkToAssignUnassing.value.studiengang +
+            " Student*in auf eine " +
+            stelle.studiengang +
+            " Stelle setzen?\n";
+    }
+
+    // Check if requested Nwk is the same
+    if (
+        stelle.namentlicheAnforderung &&
+        stelle.namentlicheAnforderung?.toUpperCase() !=
+            nwkToAssignUnassing.value.vorname.toUpperCase() +
+                " " +
+                nwkToAssignUnassing.value.nachname.toUpperCase()
+    ) {
+        warningDialogText.value +=
+            "Wollen sie wirklich " +
+            nwkToAssignUnassing.value.vorname +
+            " " +
+            nwkToAssignUnassing.value.nachname +
+            " auf diese Stelle setzen obwohl explizit " +
+            stelle.namentlicheAnforderung +
+            " angefordert wurde?\n";
+    }
+
+    // Check if Nwk is in the right semester
+    if (
+        stelle.studiengang != undefined &&
+        nwkToAssignUnassing.value.studiengang != "FISI" &&
+        stelle.studiensemester
+    ) {
+        const expectedSemester: number = +stelle.studiensemester.substring(
+            8,
+            10
+        );
+        const actualSemester = calculateSemester(nwkToAssignUnassing.value);
+        if (expectedSemester > actualSemester) {
+            warningDialogText.value +=
+                "Wollen sie wirklich eine/n Student*in im " +
+                actualSemester +
+                " Semester auf diese Stelle setzen, obwohl ein/e Student*in ab dem " +
+                expectedSemester +
+                " Semester gefordert ist?\n";
+        }
+    }
+
+    // Check if Nwk is in the right Lehrjahr
+    if (
+        stelle.ausbildungsrichtung != undefined &&
+        nwkToAssignUnassing.value.studiengang == "FISI" &&
+        stelle.ausbildungsjahr
+    ) {
+        const expectedLehrjahr: number = +stelle.ausbildungsjahr.substring(
+            4,
+            6
+        );
+        const actualLehrjahr = calculateLehrjahr(nwkToAssignUnassing.value);
+        if (expectedLehrjahr > actualLehrjahr) {
+            warningDialogText.value +=
+                "Wollen sie wirklich eine/n Auszubildende/n im " +
+                actualLehrjahr +
+                " Lehrjahr auf diese Stelle setzen, obwohl eine/n Auszubildende/n ab dem " +
+                expectedLehrjahr +
+                " Lehrjahr gefordert ist?\n";
+        }
+    }
+
+    stelleToAssignUnassign.value = stelle;
+    if (warningDialogText.value == "") {
+        assignNwk();
+    } else {
+        warningDialog.value = true;
+    }
+}
+
+function assignNwk() {
+    if (!stelleToAssignUnassign.value || !stelleToAssignUnassign.value.id) {
+        nwkToAssignUnassing.value = undefined;
+        return;
+    }
+
+    stelleToAssignUnassign.value.assignedNwk = nwkToAssignUnassing.value;
+    PraktikumsstellenService.assignNwk(
+        stelleToAssignUnassign.value.id || "",
+        stelleToAssignUnassign.value.assignedNwk?.id
+    );
+    assignedNwk.value = nwkToAssignUnassing.value;
+    emits("input", stelleToAssignUnassign.value);
+    EventBus.$emit("assignedNwk", stelleToAssignUnassign.value.assignedNwk);
+    resetWarningDialog();
+}
+
+function resetWarningDialog() {
+    warningDialogText.value = "";
+    warningDialog.value = false;
+}
+
+function calculateSemester(nwk: Nwk) {
+    if (!nwk) return -1;
+    if (nwk && nwk.studiengang == "FISI") return 0;
+    let semester: number;
+    const startYear: number = +nwk.jahrgang.substring(0, 2) + 2000;
+    const currentYear: number = new Date().getFullYear();
+    const difference = currentYear - startYear;
+    semester = difference * 2;
+    if (new Date().getMonth() > 8) semester += 1;
+    if (new Date().getMonth() < 3) semester -= 1;
+    return semester;
+}
+
+function calculateLehrjahr(nwk: Nwk) {
+    if (!nwk) return -1;
+    if (nwk.studiengang != "FISI") return 0;
+
+    let lehrjahr: number;
+    const startYear: number = +nwk.jahrgang.substring(0, 2) + 2000;
+    const currentYear: number = new Date().getFullYear();
+    lehrjahr = currentYear - startYear;
+    if (new Date().getMonth() > 8) {
+        lehrjahr += 1;
+    } else {
+        lehrjahr -= 1;
+    }
+    return lehrjahr;
+}
 
 function unassignNwk() {
     if (stelleToAssignUnassign.value?.id) {
@@ -163,6 +372,7 @@ function unassignNwk() {
             stelleToAssignUnassign.value.assignedNwk
         );
         stelleToAssignUnassign.value.assignedNwk = undefined;
+        assignedNwk.value = undefined;
     }
     resetUnassign();
 }
@@ -177,6 +387,11 @@ function resetUnassign() {
 }
 </script>
 <style scoped lang="scss">
+.custom-card-active {
+    border-color: #cfcfcf;
+    background-color: #cfcfcf;
+}
+
 .custom-card-title {
     margin-bottom: 5px;
     padding-bottom: 5px;
